ai-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
python-dotenv>=1.0.0
